Containers/Users/requirements.txt
grpcio
grpcio-tools
mysql-connector-python
redis[hiredis]